robyn_tantan/requirements.txt
anyio==3.6.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
diskimage-builder==3.31.0
flake8==5.0.4
h11==0.14.0
helpers==0.2.0
httpcore==0.16.2
httpx==0.23.3
idna==3.4
inquirerpy==0.3.4
Jinja2==3.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
multiprocess==0.70.14
nestd==0.3.1
networkx==3.2.1
paho-mqtt==1.6.1
pbr==5.11.1
pfzy==0.3.4
prisma==0.7.1
prompt-toolkit==3.0.39
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.28.1
rfc3986==1.5.0
robyn==0.49.3
rpds-py==0.10.6
six==1.16.0
sniffio==1.3.0
stevedore==5.1.0
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
uvloop==0.19.0
watchdog==2.2.1
wcwidth==0.2.6
